requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary


passlib[bcrypt]
python-dotenv
python-jose
pydantic
alembic
